Migrate sherlock265 from Qt5 to Qt6
- CMakeLists.txt: find_package Qt5 -> Qt6, update link targets
- VideoWidget.cc: remove Qt::WA_NoSystemBackground (removed in Qt6)
- VideoDecoder.cc: cast bytesPerLine() from qsizetype to int

+++ sherlock265/VideoDecoder.cc
@@ -278,7 +278,7 @@ void VideoDecoder::convert_frame_swscale(const de265_image* img, QImage & qimg)
   }
 
   uint8_t *qdata[1] = { (uint8_t *) qimg.bits() };
-  int qstride[1] = { qimg.bytesPerLine() };
+  int qstride[1] = { (int)qimg.bytesPerLine() };
   sws_scale(sws, data, stride, 0, img->get_height(), qdata, qstride);
 }
 #endif
